move layers store files to parent folder
