+++ src/modules/map/store/layers.reducer.js
