+++ src/modules/menue/components/SidePanel.js
@@ -3,7 +3,7 @@ import { BaElement } from '../../BaElement';
 import css from './sidePanel.css';
 import { closeSidePanel } from '../store/sidePanel.action';
 import { $injector } from '../../../injection';
-import { addLayer } from '../../map/store/layers/layers.action';
+import { addLayer } from '../../map/store/layers.action';
 
 
 /**
